DDF-1964: Source tab rework with publishTo registry capability. (#31)
Changes include new accordion style menus, the ability to publish to connected
registries, displaying organizational information, and numerous bug fixes such
as name display issues. Type select for a source has been removed and all configurations
are saved when editing (sources will still use an active binding as before but the actual
selection of the binding is on the source page).